Add support for l modifier on int

+++ base/my_putnbr.c
@@ -5,11 +5,12 @@
 ** my_putnbr.c
 */
 
+#include <stdint.h>
 #include <stdio.h>
 #include <unistd.h>
 #include "my.h"
 
-int my_putnbr(char *s, int n)
+int my_putnbr(char *s, intmax_t n)
 {
     int size = n < 0;
